Introduce platform project & align versions with Spring Boot

+++ src/test/java/space/npstr/wolfia/domain/ban/BanServiceTest.java
@@ -84,7 +84,7 @@ class BanServiceTest extends ApplicationTest {
         List<BanRecord> bans = this.banService.getActiveBans();
 
         assertEquals(1, bans.size());
-        assertEquals(userId, bans.get(0).getUserId());
+        assertEquals(userId, (long) bans.get(0).getUserId());
     }
 
     @Test
